+++ src/room/room.service.ts
@@ -10,7 +10,27 @@ import { Room } from './entities/room.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { UserService } from '../user/user.service';
-import { PublicRoom, RoomInfo } from '../interface';
+
+export interface RoomMetadata {
+  id: number;
+  name: string;
+  host: string;
+  user: {
+    id: string;
+    name: string;
+    isHost: boolean;
+  }[];
+}
+
+export interface PublicRoom {
+  id: number;
+  name: string;
+  host: string; // username
+  userCount: number; // count
+  vidTitle: string;
+  vidEpisode: string;
+  isLocked: boolean;
+}
 
 @Injectable()
 export class RoomService {
@@ -108,7 +128,7 @@ export class RoomService {
     return res.affected ? true : false;
   }
 
-  async roomInfo(roomId: number): Promise<RoomInfo | null> {
+  async roomMetadata(roomId: number): Promise<RoomMetadata | null> {
     try {
       const room = await this.read(roomId, ['users', 'host']);
       const userList = room.users
