+++ src/socket/socket.service.ts
@@ -8,8 +8,14 @@ import { Socket, Server } from 'socket.io';
 import { RoomService } from '../room/room.service';
 import { UserService } from '../user/user.service';
 import { WebSocketServer } from '@nestjs/websockets';
-import { Chat, Video } from '../interface';
+import { Video } from '../room/entities/room.entity';
 import { VideoService } from '../video/video.service';
+
+type Chat = {
+  senderId: string;
+  senderName: string;
+  message: string;
+};
 
 @Injectable()
 export class SocketService {
@@ -43,8 +49,8 @@ export class SocketService {
 
   async roomChanged(roomId: number) {
     this.logger.log(`roomChanged ${roomId}`);
-    const info = await this.roomService.roomInfo(roomId);
-    await this.msgInRoom(roomId, 'roomChanged', info);
+    const metadata = await this.roomService.roomMetadata(roomId);
+    await this.msgInRoom(roomId, 'roomChanged', metadata);
   }
 
   async onHostConnection(
@@ -134,7 +140,7 @@ export class SocketService {
       senderId: client.id,
       senderName: user.name,
       message,
-    } satisfies Chat;
+    };
 
     await this.msgInRoom(user.room.id, 'chat', chat);
   }
